etl/requirements.txt
# Core ETL
pyspark==3.5.1
delta-spark==3.1.0

# Azure connectivity
azure-identity>=1.16.0
azure-keyvault-secrets>=4.8.0

# Azure OpenAI / GenAI
openai>=1.30.0

# Data utilities
pandas>=2.2.0
pyarrow>=15.0.0

# SQL Server JDBC driver (download separately for Databricks)
# pyodbc>=5.1.0  # for local SQL Server connectivity

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0

# Code quality
black>=24.0.0
ruff>=0.4.0
